Add camera class and matrix passing
Matrix doesn't do much yet, and the calculations in the vertex shader are also not verified, keep that in mind.

But we do know that the matrix is passed to the GPU, using a glUniform call.

Also changed the vPos in parameter in the shader. It didn't have a specified location which was confusing, and it was still 2D. Even though we're not using the 3rd dimension yet, it' still better if it's 3D for later.